Integrate with DynamoDB for group storage
Groups are now stored both in-memory and in AWS DynamoDB.
When a group is resolved, first we look in the store, and if we can't find it, we lookup in memory. This is not because we want to cache stuff, but more because dynamo-db has some bandwidth limiations in their free account, and this allows us to work around it for the time being. Of couse it's ugly...